apps/ingestor-api/requirements/base.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings>=2.2.1
PyYAML==6.0.2
pdfplumber>=0.11.0
python-docx==1.1.0
openpyxl==3.1.2
google-api-python-client>=2.124.0
google-auth>=2.29.0
firebase-admin>=6.5.0
google-cloud-pubsub>=2.21.0
google-cloud-kms>=3.12.0
cryptography>=46.0.5
